feat: implement code splitting for recharts via dynamic imports (#948)
